crane: move inference engine into crane as engine module
Moves the continuous-batching engine, model backends, and factory out
of crane-serve into crane::engine, instead of a new crane-engine
crate.
