+++ crane/Cargo.toml
@@ -9,16 +9,19 @@ authors.workspace = true
 [dependencies]
 anyhow = "1.0.97"
 colored = "3.0.0"
-tokio = { version = "1", features = ["rt-multi-thread", "macros"] }
+tokio = { version = "1", features = ["rt-multi-thread", "macros", "sync"] }
 serde = {version = "1.0", features = ["derive"]}
 serde_json = "1.0"
 thiserror = "2.0"
+tracing = "0.1"
 nvml-wrapper = { version = "0.10", optional = true }
 
 crane-core = { path = "../crane-core" }
 hound = "3.5.1"
 ribo = "0.1.3"
 tokenizers = { workspace = true }
+candle-nn = { workspace = true }
+candle-transformers = { workspace = true }
 
 [target.'cfg(all(target_os = "macos", target_arch = "aarch64"))'.dependencies]
 candle-core = { workspace = true, features = ["metal"] }
